fix: Broken serialization due to proguard/R8 fix: Improve channel management to prevent GC on small files
