--- src-tauri/plugins/tauri-plugin-native-utils/android/src/main/java/Utils.kt
@@ -1,26 +1,0 @@
-package com.altsendme.plugin.native_utils
-
-import android.net.Uri
-import android.os.Environment
-import android.provider.DocumentsContract
-import java.io.IOException
-
-fun Uri.extractFolderOsPath(): String {
-    require(DocumentsContract.isTreeUri(this))
-
-    val path = this.path
-        ?: throw IOException("Unable to get path from selected download folder uri: $this")
-    val baseExternalPath = Environment.getExternalStorageDirectory().path
-    return try {
-        val docId = DocumentsContract.getTreeDocumentId(this)
-        val segments = docId.split(":")
-        when {
-            "primary" == segments[0] && segments.size > 1 -> "${baseExternalPath}/${segments[1]}"
-            "primary" == segments[0] -> baseExternalPath
-            segments.size > 1 -> "/storage/${segments[0]}/${segments[1]}"
-            else -> "/storage/${segments[0]}/"
-        }
-    } catch (_: Exception) {
-        path
-    }
-}
